+++ src/content-scripts/player/components/Player.vue
@@ -42,7 +42,7 @@ import { PLAYER_ACTIVITY_TIMEOUT } from '~/common/utils/constants';
 import { nextFrame } from '~/common/utils/event-loop';
 import Messenger from '~/common/utils/Messenger';
 import UsageStats from '~/common/utils/UsageStats';
-import Utils from '~/common/utils/Utils';
+import Utils from '~/common/utils/GeneralUtils';
 import { useLoadAllEpisodeData } from '../hooks/useLoadAllEpisodeData';
 import { usePlaybackRateConnector } from '../hooks/usePlaybackRateConnector';
 import { useTabUrl } from '../hooks/useTabUrl';
