+++ src/content-scripts/player/hooks/useCreateEpisodeFromThirdParty.ts
@@ -1,8 +1,8 @@
 import { RequestState } from 'vue-use-request-state';
 import { useApiClient } from '~/common/hooks/useApiClient';
+import GeneralUtils from '~/common/utils/GeneralUtils';
 import { warn } from '~/common/utils/log';
 import Mappers from '~/common/utils/mappers';
-import GeneralUtils from '~/common/utils/Utils';
 import * as Api from '~api';
 import { useUpdateEpisodeRequestState } from '../state/useEpisodeState';
 import { useDuration } from '../state/useVideoState';
