+++ src/content-scripts/player/hooks/useCreateNewTimestamp.ts
@@ -1,9 +1,9 @@
 import { useIsLoggedIn } from '~/common/state/useAuth';
 import { useGeneralPreferences } from '~/common/state/useGeneralPreferences';
 import { TIMESTAMP_TYPE_NOT_SELECTED } from '~/common/utils/constants';
+import GeneralUtils from '~/common/utils/GeneralUtils';
 import { warn } from '~/common/utils/log';
 import UsageStats from '~/common/utils/UsageStats';
-import GeneralUtils from '~/common/utils/Utils';
 import { TimestampSource } from '~api';
 import { useShowDialog, useShowLoginOverlay } from '../state/useDialogState';
 import {
