+++ src/content-scripts/player/hooks/useFindTemplate.ts
@@ -1,6 +1,7 @@
 import { useApiClient } from '~/common/hooks/useApiClient';
+import GeneralUtils from '~/common/utils/GeneralUtils';
 import { debug, log } from '~/common/utils/log';
-import GeneralUtils from '~/common/utils/Utils';
+import { sleep } from '~/common/utils/time';
 import * as Api from '~api';
 import { useEpisodeTemplate, useUpdateTemplateState } from '../state/useTemplateState';
 import { useDuration } from '../state/useVideoState';
@@ -16,7 +17,7 @@ export function useFindTemplate() {
       if (durationRef.value) {
         return durationRef.value;
       }
-      await GeneralUtils.sleep(50);
+      await sleep(50);
       debug('Waiting for player duration...');
       // eslint-disable-next-line no-constant-condition
     } while (true);
