+++ src/content-scripts/player/hooks/useLinkEpisodeUrl.ts
@@ -1,7 +1,7 @@
 import { RequestState } from 'vue-use-request-state';
 import { useApiClient } from '~/common/hooks/useApiClient';
+import GeneralUtils from '~/common/utils/GeneralUtils';
 import { warn } from '~/common/utils/log';
-import GeneralUtils from '~/common/utils/Utils';
 import * as Api from '~api';
 import { useUpdateEpisodeRequestState } from '../state/useEpisodeState';
 import { useDuration } from '../state/useVideoState';
