+++ src/content-scripts/player/hooks/useSaveTemplateTimestamps.ts
@@ -1,5 +1,5 @@
 import { useApiClient } from '~/common/hooks/useApiClient';
-import GeneralUtils from '~/common/utils/Utils';
+import GeneralUtils from '~/common/utils/GeneralUtils';
 import * as Api from '~api';
 
 export function useSaveTemplateTimestamps() {
