+++ src/content-scripts/player/hooks/useSyncTimestamps.ts
@@ -1,6 +1,6 @@
 import { useApiClient } from '~/common/hooks/useApiClient';
+import GeneralUtils from '~/common/utils/GeneralUtils';
 import { log, warn } from '~/common/utils/log';
-import GeneralUtils from '~/common/utils/Utils';
 import * as Api from '~api';
 import { useUpdateEditingState } from '../state/useEditingState';
 import { useEpisode, useEpisodeUrl, useUpdateEpisodeState } from '../state/useEpisodeState';
